requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.4
bottle==0.12.25
bottle-websocket==0.2.9
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
connexion==2.14.2
crispy==0.7.4
crispy-bootstrap4==2024.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defer==1.0.6
dirsearch==0.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-crispy-forms==2.2
djangorestframework==3.15.2
docx==0.2.4
duplicity==0.8.21
Eel==0.16.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
executing==2.0.1
fabio==2024.4.0
fasteners==0.14.1
filelock==3.15.4
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
fonttools==4.29.1
fs==2.4.12
fsspec==2024.6.1
future==0.18.2
gast==0.5.2
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gyp==0.1
h11==0.14.0
h5py==3.11.0
hdf5plugin==4.4.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.24.3
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.8.1
ntlm-auth==1.4.0
numpy==2.0.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.35.13
openpyxl==3.1.5
packaging==21.3
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.1
ply==3.11
praw==7.7.1
prawcore==2.4.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.20.0
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.2
regex==2023.12.25
reportlab==3.6.8
requests==2.25.1
requests-ntlm==1.1.0
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.8.0
seaborn==0.13.2
SecretStorage==3.3.1
silx==2.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
spark-parser==1.8.9
SQLAlchemy==2.0.2
sqlparse==0.4.4
stack-data==0.6.3
stegcracker==2.1.0
swagger_ui_bundle==1.1.0
sympy==1.13.1
systemd-python==234
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.6.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.43.3
triton==3.2.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
uncompyle6==3.9.0
unicodedata2==14.0.0
update-checker==0.18.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
whichcraft==0.6.1
xdg==5
xdis==6.0.5
xkit==0.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.2
